+++ repoze/what/middleware.py
@@ -81,7 +81,7 @@ class AuthorizationMetadata(object):
     def add_metadata(self, environ, identity):
         """
         Load the groups and permissions of the authenticated user into the
-        :mod:`repoze.who` identity.
+        :mod:`repoze.who` and :mod:`repoze.what` identity dictionaries.
         
         :param environ: The WSGI environment.
         :param identity: The :mod:`repoze.who`'s identity dictionary.
@@ -92,6 +92,12 @@ class AuthorizationMetadata(object):
         groups, permissions = self._find_groups(identity)
         identity['groups'] = groups
         identity['permissions'] = permissions
+        # Adding the groups and permissions to the repoze.what identity for
+        # forward compatibility:
+        if 'repoze.what.identity' not in environ:
+            environ['repoze.what.identity'] = {}
+        environ['repoze.what.identity']['groups'] = groups
+        environ['repoze.what.identity']['permissions'] = permissions
         # Logging
         logger and logger.info('User belongs to the following groups: %s' %
                                str(groups))